requirements.txt
# =====================================================================
# ОСНОВНЫЕ ЗАВИСИМОСТИ
# =====================================================================

# Telegram Bot API
python-telegram-bot==21.1

# Переменные окружения
python-dotenv==1.0.0

# =====================================================================
# РАЗРАБОТКА И ТЕСТИРОВАНИЕ
# =====================================================================

# Линтер (проверка кода)
pylint==3.0.0

# Форматер кода
black==24.1.0

# Type checking
mypy==1.8.0

# =====================================================================
# ДОПОЛНИТЕЛЬНЫЕ УТИЛИТЫ
# =====================================================================

# Для экспорта в CSV/Excel (если понадобится)
openpyxl==3.1.0
pandas==2.1.0

# Для работы с JSON (уже встроено, но можно явно)
# jsonschema==4.21.0

# =====================================================================
# ОПЦИОНАЛЬНО (для продвинутых фич)
# =====================================================================

# Для отправки по расписанию
APScheduler==3.10.4

# Для кэширования
cachetools==5.3.2

# Для асинхронной работы (если нужны продвинутые фичи)
aiohttp==3.9.0

# Для работы с конфигурацией YAML
PyYAML==6.0

# =====================================================================
# PRODUCTION (если деплоить)
# =====================================================================

# Для логирования в production
python-logging==0.4.9.7

# Для мониторинга ошибок (Sentry)
# sentry-sdk==1.40.0# =====================================================================
# ОСНОВНЫЕ ЗАВИСИМОСТИ
# =====================================================================

# Telegram Bot API
python-telegram-bot==21.1

# Переменные окружения
python-dotenv==1.0.0

# =====================================================================
# РАЗРАБОТКА И ТЕСТИРОВАНИЕ
# =====================================================================

# Линтер (проверка кода)
pylint==3.0.0

# Форматер кода
black==24.1.0

# Type checking
mypy==1.8.0

# =====================================================================
# ДОПОЛНИТЕЛЬНЫЕ УТИЛИТЫ
# =====================================================================

# Для экспорта в CSV/Excel (если понадобится)
openpyxl==3.1.0
pandas==2.1.0

# Для работы с JSON (уже встроено, но можно явно)
# jsonschema==4.21.0

# =====================================================================
# ОПЦИОНАЛЬНО (для продвинутых фич)
# =====================================================================

# Для отправки по расписанию
APScheduler==3.10.4

# Для кэширования
cachetools==5.3.2

# Для асинхронной работы (если нужны продвинутые фичи)
aiohttp==3.9.0

# Для работы с конфигурацией YAML
PyYAML==6.0

# =====================================================================
# PRODUCTION (если деплоить)
# =====================================================================

# Для логирования в production
python-logging==0.4.9.7

# Для мониторинга ошибок (Sentry)
# sentry-sdk==1.40.0